refactor: lift sumListCommonImpl onto SumListEmitterCommon trait (Phase 5)
Move the ~280-LOC list-shape conversion body out of
SumDispatch.sumListCommonImpl into SumListEmitterCommon.emitConvert.
The previous `gen: SumListEmitterCommon` parameter is now `this`;
helpers convertBuiltinList / hasConstantOrTypeVar / isNativeTypeVar
move to the trait alongside.

SumDispatch's six call sites collapse from
  sumListCommonImpl(new SumBuiltinListEmitter(elemRepr), input, target, pos)
to
  new SumBuiltinListEmitter(elemRepr).emitConvert(input, target, pos)

Both SumBuiltinListEmitter and SumPairBuiltinListEmitter inherit
emitConvert via the trait. SumDispatch shrinks by ~370 lines net.
