+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/SumListEmitterCommon.scala
@@ -606,6 +606,376 @@ trait SumListEmitterCommon extends SirTypeUplcGenerator {
         }
     }
 
+    /** Outbound conversions from a list-shaped value (Phase 5). The receiver
+      * `this` is the source emitter — `SumBuiltinListEmitter(elemRepr)` for
+      * cons-list sources, `SumPairBuiltinListEmitter` for pair-list sources.
+      *
+      * Body lifted from `SumDispatch.sumListCommonImpl`: the previous `gen`
+      * parameter is now `this`. Targets covered: SumBuiltinList variants,
+      * PackedSumDataList, SumPairBuiltinList variants, SumDataAssocMap, plus
+      * cross-class delegations for `SumUplcConstr` / `ProdUplcConstr` /
+      * `SumReprProxy` and the `PairIntDataList` rebuild path.
+      */
+    def emitConvert(
+        input: LoweredValue,
+        outputRepresentation: LoweredValueRepresentation,
+        pos: SIRPosition
+    )(using lctx: LoweringContext): LoweredValue = {
+        import SumCaseClassRepresentation.*
+        import ProductCaseClassRepresentation.ProdBuiltinPair
+        if input.representation == outputRepresentation then return input
+        if input.representation.isCompatibleOn(input.sirType, outputRepresentation, pos) then
+            return RepresentationProxyLoweredValue(input, outputRepresentation, pos)
+        (input.representation, outputRepresentation) match
+            // === SumBuiltinList identity ===
+            case (SumBuiltinList(inElemRepr), SumBuiltinList(outElemRepr))
+                if inElemRepr == outElemRepr =>
+                input
+            // === SumBuiltinList cross-element-repr conversions ===
+            case (SumBuiltinList(inElemRepr), SumBuiltinList(outElemRepr))
+                if hasConstantOrTypeVar(inElemRepr, outElemRepr) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                def resolveElementRepr(
+                    repr: LoweredValueRepresentation
+                ): LoweredValueRepresentation =
+                    repr match
+                        case tvr: TypeVarRepresentation =>
+                            val elemGen = lctx.typeGenerator(elemType)
+                            if !tvr.isPackedData then elemGen.defaultRepresentation(elemType)
+                            else elemGen.defaultTypeVarReperesentation(elemType)
+                        case other => other
+                val resolvedIn = resolveElementRepr(inElemRepr)
+                val resolvedOut = resolveElementRepr(outElemRepr)
+                val hasNativeTv = (inElemRepr, outElemRepr) match
+                    case (tvr: TypeVarRepresentation, _) if isNativeTypeVar(tvr) => true
+                    case (_, tvr: TypeVarRepresentation) if isNativeTypeVar(tvr) => true
+                    case _                                                       => false
+                if resolvedIn == resolvedOut
+                    || elemType == SIRType.FreeUnificator
+                    || elemType == SIRType.TypeNothing
+                    || hasNativeTv
+                then RepresentationProxyLoweredValue(input, outputRepresentation, pos)
+                else
+                    convertBuiltinList(
+                      input,
+                      elemType,
+                      resolvedIn,
+                      resolvedOut,
+                      outputRepresentation,
+                      pos
+                    )
+            // === SumBuiltinList(Constant) special cases ===
+            case (SumBuiltinList(PrimitiveRepresentation.Constant), PackedSumDataList) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                input
+                    .toRepresentation(SumBuiltinList(elemRepr), pos)
+                    .toRepresentation(PackedSumDataList, pos)
+            case (PackedSumDataList, SumBuiltinList(PrimitiveRepresentation.Constant)) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                input
+                    .toRepresentation(SumBuiltinList(elemRepr), pos)
+                    .toRepresentation(outputRepresentation, pos)
+            case (SumBuiltinList(PrimitiveRepresentation.Constant), tv @ TypeVarRepresentation(kind)) =>
+                import SIRType.TypeVarKind.*
+                kind match
+                    case Transparent =>
+                        new RepresentationProxyLoweredValue(input, tv, pos)
+                    case Unwrapped =>
+                        new RepresentationProxyLoweredValue(input, tv, pos)
+                    case Fixed =>
+                        input
+                            .toRepresentation(PackedSumDataList, pos)
+                            .toRepresentation(tv, pos)
+            case (tv @ TypeVarRepresentation(kind), SumBuiltinList(PrimitiveRepresentation.Constant)) =>
+                import SIRType.TypeVarKind.*
+                kind match
+                    case Transparent =>
+                        new RepresentationProxyLoweredValue(input, outputRepresentation, pos)
+                    case Unwrapped =>
+                        new RepresentationProxyLoweredValue(input, outputRepresentation, pos)
+                    case Fixed =>
+                        input
+                            .toRepresentation(PackedSumDataList, pos)
+                            .toRepresentation(outputRepresentation, pos)
+            // === SumBuiltinList → PackedSumDataList (via listData) ===
+            case (SumBuiltinList(_), PackedSumDataList) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val asDataList =
+                    if elemType == SIRType.TypeNothing || elemType == SIRType.FreeUnificator then
+                        input
+                    else
+                        val elemRepr =
+                            lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                        input.toRepresentation(SumBuiltinList(elemRepr), pos)
+                lvBuiltinApply(
+                  SIRBuiltins.listData,
+                  asDataList,
+                  input.sirType,
+                  PackedSumDataList,
+                  pos
+                )
+            // === SumBuiltinList catch-all (compatible element reprs) ===
+            case (SumBuiltinList(inElemRepr), out @ SumBuiltinList(outElemRepr)) =>
+                if inElemRepr == outElemRepr then input
+                else if (inElemRepr == PrimitiveRepresentation.PackedData && outElemRepr == DataData)
+                    || (inElemRepr == DataData && outElemRepr == PrimitiveRepresentation.PackedData)
+                then
+                    throw LoweringException(
+                      s"PackedData/DataData mismatch in SumBuiltinList: $inElemRepr -> $outElemRepr " +
+                          s"type=${input.sirType.show} createdEx=${input.createdEx}",
+                      pos
+                    )
+                else
+                    val elemType = retrieveElementType(input.sirType, pos)
+                    if inElemRepr.isCompatibleOn(elemType, outElemRepr, pos) then
+                        RepresentationProxyLoweredValue(input, out, pos)
+                    else convertBuiltinList(input, elemType, inElemRepr, outElemRepr, out, pos)
+            // === SumBuiltinList → SumDataAssocMap (go through SumPairBuiltinList) ===
+            case (SumBuiltinList(_), SumDataAssocMap) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val pairRepr = SumPairBuiltinList.fromElementType(elemType, pos)
+                input
+                    .toRepresentation(pairRepr, pos)
+                    .toRepresentation(SumDataAssocMap, pos)
+            // === PackedSumDataList conversions ===
+            case (PackedSumDataList, out @ SumBuiltinList(_)) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                val dataListRepr = SumBuiltinList(elemRepr)
+                val asDataList =
+                    lvBuiltinApply(SIRBuiltins.unListData, input, input.sirType, dataListRepr, pos)
+                if out == dataListRepr then asDataList
+                else asDataList.toRepresentation(out, pos)
+            case (PackedSumDataList, PackedSumDataList) =>
+                input
+            case (PackedSumDataList, SumDataAssocMap) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                val pairRepr = SumPairBuiltinList.fromElementType(elemType, pos)
+                input
+                    .toRepresentation(SumBuiltinList(elemRepr), pos)
+                    .toRepresentation(pairRepr, pos)
+                    .toRepresentation(SumDataAssocMap, pos)
+            case (PackedSumDataList, out @ SumPairBuiltinList(_, _)) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                input
+                    .toRepresentation(SumBuiltinList(elemRepr), pos)
+                    .toRepresentation(out, pos)
+            // === SumPairBuiltinList conversions ===
+            case (SumPairBuiltinList(_, _), SumPairBuiltinList(_, _)) =>
+                input
+            case (SumPairBuiltinList(_, _), SumDataAssocMap) =>
+                lvBuiltinApply(SIRBuiltins.mapData, input, input.sirType, SumDataAssocMap, pos)
+            case (SumPairBuiltinList(_, _), PackedSumDataList) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val dataElemRepr =
+                    lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
+                input
+                    .toRepresentation(SumBuiltinList(dataElemRepr), pos)
+                    .toRepresentation(PackedSumDataList, pos)
+            case (
+                  SumPairBuiltinList(inKeyRepr, inValueRepr),
+                  out @ SumBuiltinList(outElemRepr)
+                ) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val inElemRepr = ProdBuiltinPair(inKeyRepr, inValueRepr)
+                convertBuiltinList(input, elemType, inElemRepr, outElemRepr, out, pos)
+            case (SumBuiltinList(inElemRepr), out @ SumPairBuiltinList(outKeyRepr, outValueRepr)) =>
+                if isNilType(input.sirType) then lvPairDataNil(pos, input.sirType, out)
+                else
+                    val elemType = retrieveElementType(input.sirType, pos)
+                    val outElemRepr = ProdBuiltinPair(outKeyRepr, outValueRepr)
+                    convertBuiltinList(input, elemType, inElemRepr, outElemRepr, out, pos)
+            // === SumDataAssocMap conversions ===
+            case (SumDataAssocMap, SumPairBuiltinList(_, _)) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val pairRepr = SumPairBuiltinList.fromElementType(elemType, pos)
+                lvBuiltinApply(SIRBuiltins.unMapData, input, input.sirType, pairRepr, pos)
+            case (SumDataAssocMap, _) =>
+                val elemType = retrieveElementType(input.sirType, pos)
+                val pairRepr = SumPairBuiltinList.fromElementType(elemType, pos)
+                input
+                    .toRepresentation(pairRepr, pos)
+                    .toRepresentation(outputRepresentation, pos)
+            // === PairIntDataList for list-typed values (suspicious — DataConstr should
+            //     not exist for List types per design, but PairIntDataList is its
+            //     unpacked form. Probe with logging if hit during refactor).
+            case (PairIntDataList, SumBuiltinList(elementRepr)) =>
+                throw LoweringException(
+                  s"PairIntDataList → SumBuiltinList($elementRepr) for ${input.sirType.show}: " +
+                      "DataConstr should not exist for List types.",
+                  pos
+                )
+            case (PairIntDataList, _) =>
+                // Reconstruct DataConstr from (tag, fieldList), then route via PackedSumDataList.
+                val asDataConstr = lvBuiltinApply2(
+                  SIRBuiltins.constrData,
+                  lvBuiltinApply(
+                    SIRBuiltins.fstPair,
+                    input,
+                    SIRType.Integer,
+                    PrimitiveRepresentation.Constant,
+                    pos
+                  ),
+                  lvBuiltinApply(
+                    SIRBuiltins.sndPair,
+                    input,
+                    SIRType.List(SIRType.Data.tp),
+                    SumBuiltinList(DataData),
+                    pos
+                  ),
+                  input.sirType,
+                  DataConstr,
+                  pos
+                )
+                asDataConstr
+                    .toRepresentation(PackedSumDataList, pos)
+                    .toRepresentation(outputRepresentation, pos)
+            // === TypeVarRepresentation ===
+            case (_, tv: TypeVarRepresentation) =>
+                import SIRType.TypeVarKind.*
+                tv.kind match
+                    case Transparent =>
+                        new RepresentationProxyLoweredValue(input, tv, pos)
+                    case Unwrapped =>
+                        val targetUnderlying = defaultRepresentation(input.sirType)
+                        val converted = input.toRepresentation(targetUnderlying, pos)
+                        new RepresentationProxyLoweredValue(converted, tv, pos)
+                    case Fixed =>
+                        val inputAsData = input.toRepresentation(PackedSumDataList, pos)
+                        new RepresentationProxyLoweredValue(inputAsData, tv, pos)
+            case (tv: TypeVarRepresentation, _) =>
+                import SIRType.TypeVarKind.*
+                tv.kind match
+                    case Transparent =>
+                        new RepresentationProxyLoweredValue(input, outputRepresentation, pos)
+                    case Unwrapped =>
+                        val sourceUnderlying = defaultRepresentation(input.sirType)
+                        val r0 = new RepresentationProxyLoweredValue(input, sourceUnderlying, pos)
+                        r0.toRepresentation(outputRepresentation, pos)
+                    case Fixed =>
+                        if input.representation == outputRepresentation then input
+                        else
+                            val r0 = new RepresentationProxyLoweredValue(
+                              input,
+                              PackedSumDataList,
+                              pos
+                            )
+                            r0.toRepresentation(outputRepresentation, pos)
+            // SumReprProxy: unwrap and delegate
+            case (_: SumReprProxy, _) =>
+                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
+            case (_, _: SumReprProxy) =>
+                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
+            // SumUplcConstr → anything: delegate
+            case (_: SumUplcConstr, _) =>
+                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
+            // anything → SumUplcConstr: delegate
+            case (_, _: SumUplcConstr) =>
+                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
+            // ProdUplcConstr value at a sum-list-type site: wrap as singleton SumUplcConstr
+            case (puc: ProductCaseClassRepresentation.ProdUplcConstr, _) =>
+                val wrappedRepr =
+                    SumUplcConstr(scala.collection.immutable.SortedMap(puc.tag -> puc))
+                val wrapped = new RepresentationProxyLoweredValue(input, wrappedRepr, pos)
+                SumUplcConstrSirTypeGenerator.toRepresentation(
+                  wrapped,
+                  outputRepresentation,
+                  pos
+                )
+            case _ =>
+                throw LoweringException(
+                  s"Unexpected representation conversion for ${input.sirType.show} from ${input.representation} to $outputRepresentation",
+                  pos
+                )
+    }
+
+    /** Cross-element-repr conversion for BuiltinList, applied via ScalusRuntime.mapList. */
+    protected def convertBuiltinList(
+        input: LoweredValue,
+        elemType: SIRType,
+        resolvedIn: LoweredValueRepresentation,
+        resolvedOut: LoweredValueRepresentation,
+        outputRepresentation: LoweredValueRepresentation,
+        pos: SIRPosition
+    )(using lctx: LoweringContext): LoweredValue = {
+        import SumCaseClassRepresentation.*
+        val inElemUplcType = resolvedIn.uplcType(elemType)
+        val outElemUplcType = resolvedOut.uplcType(elemType)
+        val inListType = SIRType.BuiltinList(inElemUplcType)
+        val outListType = SIRType.BuiltinList(outElemUplcType)
+        val isPairIn = resolvedIn match
+            case _: ProductCaseClassRepresentation.ProdBuiltinPair => true
+            case _                                                 => false
+        val isPairOut = resolvedOut match
+            case _: ProductCaseClassRepresentation.ProdBuiltinPair => true
+            case _                                                 => false
+        val inListRepr: LoweredValueRepresentation =
+            if isPairIn then SumPairBuiltinList.fromElementType(elemType, pos)
+            else SumBuiltinList(resolvedIn)
+        val outListRepr: LoweredValueRepresentation =
+            if isPairOut then SumPairBuiltinList.fromElementType(elemType, pos)
+            else SumBuiltinList(resolvedOut)
+        val outNil =
+            if isPairOut then SumPairBuiltinListEmitter.genNil(outListType, pos)
+            else new SumBuiltinListEmitter(resolvedOut).genNil(outListType, pos)
+        val convFn = lvLamAbs(
+          "elem",
+          elemType,
+          resolvedIn,
+          elem => elem.toRepresentation(resolvedOut, pos),
+          pos
+        )
+        val afterFnType = outListType ->: inListType ->: outListType
+        val afterFnRepr = LambdaRepresentation(
+          afterFnType,
+          InOutRepresentationPair(
+            outListRepr,
+            LambdaRepresentation(
+              inListType ->: outListType,
+              InOutRepresentationPair(inListRepr, outListRepr)
+            )
+          )
+        )
+        val afterNilType = inListType ->: outListType
+        val afterNilRepr =
+            LambdaRepresentation(afterNilType, InOutRepresentationPair(inListRepr, outListRepr))
+
+        val result = lvApplyDirect(
+          lvApplyDirect(
+            lvApplyDirect(ScalusRuntime.mapList, convFn, afterFnType, afterFnRepr, pos),
+            outNil,
+            afterNilType,
+            afterNilRepr,
+            pos
+          ),
+          input,
+          input.sirType,
+          outListRepr,
+          pos
+        )
+        if outputRepresentation == outListRepr then result
+        else RepresentationProxyLoweredValue(result, outputRepresentation, pos)
+    }
+
+    private def hasConstantOrTypeVar(
+        a: LoweredValueRepresentation,
+        b: LoweredValueRepresentation
+    ): Boolean =
+        a == PrimitiveRepresentation.Constant || b == PrimitiveRepresentation.Constant ||
+            a.isInstanceOf[TypeVarRepresentation] || b.isInstanceOf[TypeVarRepresentation]
+
+    private def isNativeTypeVar(tvr: TypeVarRepresentation)(using
+        lctx: LoweringContext
+    ): Boolean =
+        import SIRType.TypeVarKind.*
+        tvr.kind match
+            case Transparent | Unwrapped => true
+            case Fixed                   => false
+
 }
 
 object SumListEmitterCommon {
